Move getPendingEntities query to dashboard container
