+++ src/pages/Admin/Home/Dashboard/component.tsx
@@ -1,4 +1,4 @@
-import React, { FunctionComponent, useState } from "react";
+import React, { FunctionComponent } from "react";
 import { Window } from "$components/Window";
 import { Menu } from "./Menu";
 import { TaskDetail } from "./TaskDetail";
@@ -6,26 +6,35 @@ import { TaskList } from "./TaskList";
 import styles from "./styles.module.scss";
 import { IApprovable, IApprovableFilter } from "$interfaces/Approvable";
 
-const Dashboard: FunctionComponent = () => {
-  const [selectedTask, setSelectedTask] = useState<IApprovable>();
-  const [filter, setFilter] = useState<IApprovableFilter>({});
+export const Dashboard: FunctionComponent<IDashboardProps> = (
+  {
+    approvableEntities,
+    selectedTask,
+    setSelectedTask,
+    filter,
+    setFilter
+  }
+) => (
+  <Window width="fullWidth">
+    <div className={styles.mainContent}>
+      <Menu filter={filter} onSelectFilter={setFilter}/>
+      <TaskList
+        approvableEntities={approvableEntities}
+        selectedTask={selectedTask}
+        onSelectTask={setSelectedTask}
+      />
+      <TaskDetail
+        selectedTask={selectedTask}
+        onStatusUpdate={() => setSelectedTask(undefined)}
+      />
+    </div>
+  </Window>
+);
 
-  return (
-    <Window width="fullWidth">
-      <div className={styles.mainContent}>
-        <Menu filter={filter} onSelectFilter={setFilter}/>
-        <TaskList
-          filter={filter}
-          selectedTask={selectedTask}
-          onSelectTask={setSelectedTask}
-        />
-        <TaskDetail
-          selectedTask={selectedTask}
-          onStatusUpdate={() => setSelectedTask(undefined)}
-        />
-      </div>
-    </Window>
-  );
-};
-
-export { Dashboard };
+interface IDashboardProps {
+  selectedTask?: IApprovable;
+  setSelectedTask: (task?: IApprovable) => void;
+  filter: IApprovableFilter;
+  setFilter: (filter: IApprovableFilter) => void;
+  approvableEntities: IApprovable[];
+}
